UI should display the feed once form is submitted
