+++ app/ui/components/Feed/index.js
@@ -1,0 +1,1 @@
+export { Feed } from "./Feed";
