+++ app/ui/components/TopicsForm/index.js
@@ -1,0 +1,1 @@
+export { TopicsForm } from "./TopicsForm";
