+++ app/ui/components/index.js
@@ -1,1 +1,3 @@
 export { App } from "./App";
+export { Feed } from "./Feed";
+export { TopicsForm } from "./TopicsForm";
