+++ app/ui/hooks/index.js
@@ -1,0 +1,1 @@
+export { useFeedSubscription } from "./useFeedSubscription";
